--- qiskit/providers/ibmq/api/rest/version_finder.py
@@ -1,55 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2018, 2019.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""Version REST adapter."""
-
-from json import JSONDecodeError
-from typing import Dict, Union
-
-from .base import RestAdapterBase
-
-
-class VersionFinder(RestAdapterBase):
-    """Rest adapter for the version finder."""
-
-    URL_MAP = {
-        'version': '/version'
-    }
-
-    def version(self) -> Dict[str, Union[str, bool]]:
-        """Return the version information.
-
-        Returns:
-            A dictionary with information about the API version,
-            with the following keys:
-
-                * ``new_api`` (bool): Whether the new API is being used
-
-            And the following optional keys:
-
-                * ``api-*`` (str): The versions of each individual API component
-        """
-        url = self.get_url('version')
-        response = self.session.get(url)
-
-        try:
-            version_info = response.json()
-            version_info['new_api'] = True
-        except JSONDecodeError:
-            return {
-                'new_api': False,
-                'api': response.text
-            }
-
-        return version_info
