+++ qiskit/providers/ibmq/ibmqfactory.py
@@ -486,3 +486,4 @@ class IBMQFactory:
                 # Catch-all for errors instantiating the provider.
                 logger.warning('Unable to instantiate provider for %s: %s',
                                hub_info, ex)
+                raise
